requirements.txt
networkx==3.2.1
pillow==11.1.0
numpy==1.23.5
PyYAML==6.0.2
openai==1.64.0
argparse==1.4.0
torch==2.5.0
torchvision>=0.15.0
opencv-python>=4.7.0
segment-anything
requests>=2.31.0
matplotlib>=3.2.2
scipy>=1.4.1
tqdm>=4.41.0
protobuf<4.21.3
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
ipython==8.12.3
psutil==5.9.8
thop==0.1.1.post2209072238
